update cli reporting, fix repo errors, add volatility

+++ day-15-stock-sentiment/src/models.py
@@ -17,7 +17,7 @@ class ScoredNews:
     headline: str
     source: str
     published_at: datetime
-    score: int
+    score: float
 
 
 @dataclass
